Added icons to the sidebar, which are royalty free icons I downloaded from online.  Also the sidebar list items are fully link clickable and not just the text.
To do this I done in css:
a{
display:block;
}

And then to stop an underline from showing underneath the text I added this:
a, u {
    text-decoration: none !important;
}

Both client and advisor sessions now have icons in their sidebar.  When on a page, the icon in the sidebar is set to green. If the icon is green, you are on that page.